It is really great when you learn to read :)
I had totally misunderstood the java.net.URL javadocs, so this wasn't working at all.  This shouldn't effect any outstanding bugs (due to a weird combination of bugs in the previous version, that prevented this class from doing much of anything).  However, it will now hopefully work as advertised.

TODO: make the loading of this configurable.