As each request passes through the controller servlet, call the new Request.selectApplication() utility method.  This method will identify the appropriate sub-application by matching prefixes, and then expose two request attributes that are used for the remainder of the processing:
* The ApplicationConfig object for the current application is exposed
  under key Action.APPLICATION_KEY

* The MessageResources object for the current application is exposed
  under key Action.MESSAGES_KEY

Now, the remaining logic can be modified to consult these two request
attributes to make decisions that depend on the current sub-application.

NOTE -- This processing is NOT performed on direct requests for JSP pages --
it is only done on requests through the controller.  In a Servlet 2.3
environment we will be able to deal with that by moving this logic to a
filter that is mapped to all requests.  In the mean time, applications that
will operate in a multi-app controller environment MUST pass all of their
requests through the controller servlet to work properly.

NOTE -- I am assuming that data sources are still accessed directly by their
servlet context attribute keys, and that all data source key names are unique
across sub-applications.