Tighten up the definitions in the DTD, including using a new entity called "AttributeName" that will be used instead of BeanName for attributes that do not need to conform to Java variable name requirements.  (Suggested by Donnie Hale).
Make the DTD and the Digester rules consistent with respect to when the
"className" attribute is used to override the configuration bean
implementation class, and when a nested <set-property> value is supported.

Add support for a new <form-property> element that can be nested inside
<form-bean>.  This will be used to add support for creating ActionForm beans
without having to actually create all the classes individually.