Initial check-in of dynamic ActionForm support (uses the new DynaBean support in the Commons version of BeanUtils).  Now, you have three choices for how to implement ActionForms:
* Implement your own ActionForm subclass with standard JavaBean
  properties, as before.

* Use the standard class "org.apache.struts.action.DynaActionForm",
  which will be populated with properties defined by <form-property>
  elements nested inside the <form-bean> element.  Note that you can
  also set initial values in the <form-property> element; these are
  used in the default implementation of reset().  Intelligent defaults
  take care of the most common cases.  Note that the default validate()
  method does NO checking, so you will need to do everything in your
  Action.

* Implement a subclass of "org.apache.struts.action.DynaActionForm"
  if you need to customize just the reset() and/or validate() methods.
  In this case, you will need to tell Struts that you still want
  dynamic bean support:
    <form-bean name="logonForm" dynamic="true"
               type="com.mycompany.mypackage.MyDynaActionForm">
      ... <form-property> elements ...
    </form-bean>

Just as an illustration of how this is used, the logon form bean in the
Struts example app has been replaced by a corresponding dynamic bean.
Note that I had to adjust the Action to use PropertyUtils to read the
properties, in order to make it not care about the actual implementation --
this approach will be common when rapidly prototyping an applicaiton, and
then going back and implementing specific ActionForm classes when needed
later.  PropertyUtils knows how to handle either kind of ActionForm
transparently, so no changes to the Action are required.