Follow Martin's suggestion and allow more than one <message-resources> element in the struts-config.xml file.  The default (as you might expect :-) is backwards compatible.
Also, refine the storing of data sources and message resources into servlet
context attributes.  The actual attribute used is the specified key plus the
sub-application prefix (which, for the default sub-app, is again backwards
compatible).  The net effect of this is that data source and message
resources keys do *not* have to be globally unique.