removed the SelectModule action.  Security is enforced and the acl is built based on the current module as it exists before the SelectModule action is run.  Allowing someone to change modules after these actions have run is a potential security hole.
We could perform any change of modules within the session validator, however,
I went with the idea started by Maarten and the current module is always
determined by the request parameter.

The method that allows a template to specify a link which will change the
module is left protected and only made public for a special link tool that
is added to the context for the SelectModule.vm template.

I went this route because modules are a likely point of integration with
external systems so it is best if the ability to switch modules is localized.
Of course the UI will probably change to using a form to switch modules as soon as I do this.  If so the session validator will need to be altered.


Issue number:  136
Obtained from:
Submitted by:
Reviewed by: