errors like this were reported on the list...
compile:
    [javac] Compiling 307 source files to
/usr/local/scarab/target/webapps/scarab/WEB-INF/classes
    [javac]
/usr/local/scarab/target/webapps/scarab/WEB-INF/src/org/tigris/scarab/om/Rep
ort.java:110: Blank final variable 'ACT_ATTRIBUTE_ID' may not have been
initialized. It must be assigned a value in an initializer, or in every
constructor.
    [javac]     private static final String ACT_ATTRIBUTE_ID;
    [javac]                                                 ^

I believe he was using javac from JDK 1.2.2 on RHAT...this should fix those
errors. apparently this is a 1.2.2 compiler bug.

the way that john wrote the code was the right way, but i would rather not
deal with having to support people who are compiling with that javac...so,
this should fix those errors.

-jon