the order should be:
1.  Load any project specific overrides
2.  Load any global properties
3.  Load defaults

This allows one to specify xerces.jar=1.4.3 in
~/build.properties for the 5 projects a user is
working on.  Then when one of the projects moves
to 2.0.0, they can put a <project1>/build.properties
for that specific property.

With the current setup the user would be forced to
remove the property from their ~/build.properties and
add a build.properties for all projects.