Added an ajp12 option to go with the new ajp13 option and reworked the logic a bit.
If either ajp12 or ajp13 is set (or the last one if both are set), that protocol
will be used.  If an ajpid file is not specified, the associated default will be
used.

If neither ajp12 or ajp13 is specified, and no ajpid file is specified, then
ajp13 is tried first with failover to ajp12 if the ajp13 file is not found or
shutdown isn't enabled.  If an ajpid file is specified, it defaults to using
ajp12.

Updated the help output.