Caching the permission results internally.  The results probably need to be stored externally to allow a mechanism to invalidate the cache.  Under the current scheme an admin cannot modify the status of a logged in user.
Some sort of caching is called for as this example took the processing time of
ViewIssue.vm from 3 seconds to 1 second.  Even though it is not a perfect
solution, I am leaving it in place as I am trying to evaluate other sql
inefficiencies and without this cache permission related sql will dominate
the logs.