this is a bit of a hack. we don't use the TURBINE_GROUP table, instead, we use the SCARAB_MODULE table because TURBINE_GROUP == SCARAB_MODULE.
now, when we import data into the database, the foreign key references
for the other TURBINE_* tables require implicit constraint integrity because
of the FK reference to TURBINE_GROUP.GROUP_ID. if we remove that, then
the integrity is not needed. we aren't using the TURBINE_GROUP table anyway
so this works around that problem and the data can be imported into
databases like oracle/postgresql that require key integrity.

-jon