xslt from scarab-schema.xml to html provided by Alan Hodgkinson <alan@softxs.ch>
I added a build target html-schema to generate the doc.  I had some problems
with the xalan/xerces jars, so decided to keep things simple for now and use
a custom classpath for the target.  Note: we must use xerces-1.4.4 for the
application as tomcat uses it (at least for scarab with sourcecast).

Added the new doc to www directory and added a link to it.