moved a fix into the fixes section.
added the issue id's to elicia's issues so that she can see an example
of what i would like her to do. i want to see the issue id in not only
the cvs log entry but also the changes file. i was not clear enough on
this earlier and apparently the existing entries with the issue id noted
were not enough of an example.

-jon