it compiles, but it isn't fully fleshed out or working yet. checking in now just to prevent any loss. this is going to take some time to fully go through each of these classes and test/clean stuff up. i will to it in steps and checkin as each step is complete.
right now, it can properly parse the <module> but pukes when it
goes into parsing the user tags...mostly as a result of the change
from using digester.push/pop() to using the ImportBean to store
data. this will get fixed as i clean stuff up...

i also added a <description> element to the <module> element.

-jon