another round of changes...
it is now possible to run the ant -f import.xml db-validate on a
user .xml file and it will process through and create the right objects
without throwing exceptions anymore.

more work is still needed...next will be actually doing a db-insert to
create/update the users and testing to make sure that all that works
as it should.

also need to optimize the code to make sure that any duplicate users
are not processed in the code...

getting closer...

-jon