ok, this change will probably piss the UI guys off (/me notices todd's head pop up above the cubicle)...
anyway, what it does is add focus around the 'sections' within the
left hand nav. however, it doesn't do it conditionally as I might suspect
that todd would want (ie: if in the admin, focus on admin. if in issue types,
focus on issue types). implementation of that would require significantly
more time/code to implement and probably would not be worth it in the long
run.

however, i do feel that adding focus to the boxes all the time serves
another UI purpose which is to add seperation to the various sections
of navigation. it also provides a more 'boxy' feel to the navigation
in order to match the style of the box for the help text.

comments welcome. :-)

-jon