Another pass at move/copy issue:
Moved some code from action to Issue business object.

Added permission check : for Copy Issue, it's ISSUE | ENTER ; for Move Issue, ISSUE | ENTER for destination module, and MODULE | EDIT for source module.

Adding user attributes as a comment for now.

Set committedBy for the copied issue to be the same as for the original.

Testing that the issue type exists in the destination module.