Removed all uses of FinderResults and replaced them with a seperate Collection and ReadAhead object.
Fixed several bugs that caused an entity to loose it's readahead info.

Changed code to allow entity to be associated with a singleton list and
readahead none.  This way a readahead none query can stop old read ahead
info from causeing a readahead on-load.

Remove use of LRUCache because it is overkill, and hard to understand.