Added new true-mapping and false-mapping elements to the type mappings.
Most are mapped to 1 and 0 respecively, because I don't know the correct
mappings for all databases.  We will have to wait for user imput.

The mapping for Hypersonic is (1=1) and (1=0) because it has problems
parsing the literal values TRUE and FALSE is some where clauses.  When
they fix the parser (1.7?) it should be changed back to TRUE and FALSE for
better readability.