Major simplification of code.
Added support for char and Character to be loaded and stored via a String
and VARCHAR.  This should fix Oracle's problems with characters.

Added better support for CLOB.  CLOBs will be set using setCharacterStream.
This will not fix Oracle's problems with large streams, but makes CLOBs
usable for most databases.