Experimental support for 2.3 style mapped-welcome files.
This of off be default, and probably uses too many recorces if you have more than a few welcome-files.  What it enables is JspC compilations and declarations like:

<servlet-mapping>
  <servlet-name>WelcomeServlet</servlet-name>
  <url-pattern>/foobar/</url-pattern>
</servlet-mapping>

For now, should be considered experimental only.  However, I'd be interested to hear how much overhead it adds if anyone wants to test.