DynaActionFormClass - When a DynaActionForm instance is first created, initialize all properties to the initial values configured in the <form-property> elements, similar to what calling the default reset() method does.
Flesh out tests for DynaActionFormClass, add initial test for
DynaActionForm.

FIXME - Deal with the static collections in DynaActionFormClass when
struts.jar is in a shared class loader.

FIXME - Support a syntax for initializing array, indexed, list, and
mapped type properties in the "initial" attribute of <form-property>
elements, so that such properties can be preset with rational values.