It is now possible to initialize the value of an indexed property in a DynaActionForm bean, with syntax like this:
  <form-property name="intArray" type="int[]"
              initial="{ 0, 1, 2, 3}"/>

  <form-property name="doubleArray" type="double[]"
              initial="0.1 2.3 4.5"/>

  <form-property name="stringArray", type="java.lang.String[]",
              initial="'String 0', 'String 1', 'String 2'"/>

if you are using a version of commons-beanutils >= the 20020630 nightly build
(or any upcoming 1.4 release).  In addition, you can register your own
String->ArrayType converters in ConvertUtils for arrays of non-primitive
types, and parse the initial value string any way you like.

The syntax for primitive and String indexed initialization was designed to be
similar to that a Java developer uses to initialize arrays directly:

  int intArray[] = new int[] { 0, 1, 2, 3};

with the following adjustments:

* Only literal values are accepted -- no expression evaluation the way
  a compiler would do it.

* Beginning and ending "{" and "}" are optional -- you need both or neither.

* Commas in between elements are considered whitespace, so they are
  optional as well.

* You can use either single quotes or double quotes around strings, and
  embed the opposite type inside the String.  (Commas can, of course,
  be embedded as well.

When the reset() method of a DynaActionFormBean is called (including when it
is created initially for you), the default functionality is to initialize
all form bean properties to the initial values again.  You can override this
in your own DynaActionForm subclass if need something extra.