Support the new "inputForward" boolean property on ControllerConfig that (if set to true) defines the "input" attribute of an <action> element to be the name of a local or global forward that is used to calculate the ultimate URL to which control is returned on validation errors.  The default behavior (false) assumes that the "input" attribute is a subapp-relative URL (backwards compatible with the previous behavior).
ActionMapping - Add a new getInputForward() method that will return an
ActionForward corresponding to the input form, no matter whether or not
the "inputForward" property is set.  This is useful in Actions that want
to create "validation error" messages and return to the input form
themselves, such as LogonAction in the struts-example webapp.

struts-config.xml - The struts-example webapp now sets the "inputForward"
property to true, and utilizes the new features.