refactored the Activity creation code into the Manager and created individual methods for creating specific activities...there is still more work to do here to try to centralize the creation of Activities and Transactions, but it is a start. this is so that we can eventually have a single system for logging events in Scarab because there is web actions which create events and there is soon to be xml actions which create events...if we have to duplicate that code for each action (web, xml, xml-rpc???), it is going to be terribly messy.
the other issue that needs to be fixed is that the way that events are
logged in scarab...it is done without much consideration to what
needs to be logged and where...we need to re-consider the current
system and clean it up...that is where my next set of actions is
going to occur. this will involve moving a lot of code out of the
web actions into the business objects (where it belonged in the first place)
so that the xml actions can take advantage of a single code base
and single api. it will then be more easily possible to create
other actions like rpc actions on top of this code base.

-jon