User searches had a couple errors.
if the user involved in a search on user attributes is not associated with any issues, the result should be no issues.  However the result is all issues which would be returned ignoring that search criteria are returned.  If the user is associated with at least one issue, the correct result is returned.

Also if a search is for "Created By" only, all issues created by the user will be returned without regard for the modules and issue types that should limit the results.

Issue Number: SCB34
http://scarab.whichever.com:8081/issue/id/SCB34