Hi all,
Attached is a small patch for SearchNav.vm.  Currently, we have a couple
modules that only have only one issue type in them.  However, they keep
clicking on the dropdown for the issue types at the top, and are confused
when nothing shows up (because there is only one module).  Because the first
module is used as the default, it seems that the issue type dropdown should
only show if there are more then one issue.

Eric

patch from EPugh@upstate.com