removed deprecated code in DependType
updated modifyissue to do permission checking (need at least issue | edit)
to modify an issue

more re-factoring of code (this time, it was a near re-write) so that
the business logic is removed from the action and placed into the Issue
om object. I also did some work on the way that activities were being
created as well as the messages displayed to the user.

-jon