Issues: 240, 284, 425, 699, 1045
Made ArgoUML use predecessor and activator to automatically calculate
sequence numbers.
Made the factory set predecessor and activator as good as possible.
Made "in diagram" editing of messages on collaboration diagrams conform
better to the UML specification.
Made "in diagram" editing of messages on collaboration diagrams create
operations if the receiver is properly specified and has no operation by that
name with a proper number of parameters.