removed several of the From email address properties.  In some cases the addresses were not even being used, as the modifyissue address was used in most cases.
For emails that occur in the context of a module there can be a per module
address used as the ReplyTo address (and if no other user is applicable it
is used as the From as well).  The implementation currently works like:

 If a scarab.email.default.fromName/Address property exists it is used,
otherwise if the module has an archive email address defined it will use
that.

It would be good to add a new attribute to modules for this specifically
and then the order could be reversed.