Submitted by:  Thierry Lach
Patches to get datadump, datadtd, and datasql executing.

Datadump does not appear to work correctly for Oracle, but this appears to be a torque issue (null column values seem to repeat the data from the previous column).

Datasql runs but with the data I was using encounters an out-of-memory abort.

But prior to this change, they didn't even execute.