From: Jean-François El Fouly <jean-francois.elfouly@wanadoo.fr>
I suspect there is an erroneous carriage return in ScarabBundle_en_US.properties, line 1410 (after the key UQT16).

I also note in this help entry that the original version said that &amp;&amp; could be used instead of AND in a text query. In the new version, there is just one &.
Who's right ?

&amp;&amp;