Previous algorithm for sorting by attribute value was: 1.  Get list of issues from the db without sorting. 2.  Go through list and separate into lists of issues that have a value for     the attribute and those that do not. 3.  Get a list of sorted issues from the db for issues that have the     attribute set 4.  Append the list of issues that did not have the value set.
Step #2 was very inefficient and though I did not measure it, it appeared as
though it would be about 1-10 hours to complete to sort 20000 issues of which
all had the attribute set (so it would have been best to avoid #2 all together.)

The method has been rewritten to
1. Get the count of total issues without any sorting.
2. Get a list of sorted issues from the db for issues that have the
   attribute set.
3. Compare 1 and 2.  If all issues had a value for the attribute we are done.

otherwise
4. Get list of issues from the db that don't have the value set.

#4 is done through a potentially large NOT IN sql list to exclude the issues in
#2.  I tested a reasonably bad case with a total result set of 20000 issues with
 1 not having a value for the sort attribute so the NOT IN list contained
19999 ids. Mysql was able to handle the sql and the result was not slower than
than getting the sorted list.

Total query time to get the ids ~2 secs.  Though I still need to commit an
index change to achieve that.