Proof-reading / formatting / spell-check pass. No significant content changes.
Most of the changes are line-wraps from applying the coding rules to our XML:

* Line wraps at column 78.
* Sentences ("statements") begin a new line.

Starting a new line at each sentence minimizes future changes, since the rest of the paragraph does not need to be re-wrapped.
It also helps highlight longer sentences (a readability issue).

Other conventions:

* For paragraph and paragraph-like elements (UL, LI, TD), the element is placed above and below the text to increase legibility.

* Single space between paragraph-like elements, unless there is only one element in the block.

I apologize if this annoys anyone, but the XML formatting is getting out of control
and it's sometimes difficult to make routine changes.

I didn't use a specific tool to do this,but just applied the rules during an overall review of the documentation.