fixed file upload on wizard3...the problem was that the issue hasn't been created yet, so we can't just assume that we can immediately save the file. this is somewhat of an edge case.
store the attachments in a temporary ArrayList (sync'd the methods, not
sure if it is really needed, but can't hurt)

elicia: note how i changed your code. there was no reason to re-associate
the attachment or call the save() method again because it is already
done in ActivityManager.create()...please check out that code so you
have an understanding of it.