Modified Files: 	src/java/org/tigris/scarab/tools/ScarabRequestTool.java 	src/templates/navigations/SearchNav.vm ---------------------------------------------------------------------- Issue number: iz#14186
ScarabRequestTool.java - since these methods get queries of only one
scope, there is no reason to sort by scope.  Sorting the queries by name.
This removes a join so it is a database performance enhancement.

SearchNav.vm - several performance improvements and a bug fix.  variables were
being created and not used or created for one case when it could be used
many places. Also the $currentIssueType variable did not exist and so the
template dropdown was never shown.