a bunch of long standing FIXME internal api changes that needed to happen now because i need to use some of these methods for the xml import process. namely, i need to overload the module.getAttributes() method to return a list of attributes in the module.
what i did was get rid of using Object[] and use List instead. this is
a cleaner way to do things since one will get the benefit of the collections
api.

this should not have any effect on stuff since it was a fairly simple
api change, it just happened to touch a lot of files.

-jon