Make the Cactus tests run successfully under Tomcat 4.0.6, when running the "test.tomcat.40" target.
The problem is a bug in Tomcat 4.0.6, where the JSP page compiler incorrectly
does an implicit "import org.apache.jasper.runtime.*" in the generated code
for each page.  Unfortunately, that makes the class name "JspException"
ambiguous.  I corrected this by fully qualifying the exception class we
really want (javax.servlet.jsp.JspException), which runs on 4.0 and 4.1.

I have't tried these on 3.3 yet -- that will be later.