forward port of changes on the SCAST_T1 branch.  The merge was performed  by tagging the current state of the SCAST_T1 branch as SCAST_T1_MERGE_PT_3,  then on a HEAD working copy running the command  cvs up -kk -j SCAST_T1_MERGE_PT_2 -j SCAST_T1_MERGE_PT_3
In the cases where there were conflicts that needed to be resolved, I
took the code from the SCAST_T1 branch as being the definitive code to
use.

Note that I had to manually add a couple of keys (DateCommitted,
CommittedBy) from the ScarabBundle_en_US.props file into the one in HEAD
for some reason.