Fixed bug where an ejbStore invokes a finder caused an infinate loop.
Fixed bug where an ejbStore modifies another entity the modification was
not reconized.

Major preformance increase.  Now only the entity on the head of the
invocation stack is considered always dirty.  Entities are put in the
dirty set before they invoke another entity and when they leave the
invocation (assuming they are in a transaction).

Added better support for ReadOnly.

Added better support for non-transactional entities.