Adds the "validwhen" validation rule.
Yes, this is late in the game for a new feature, but I'm commiting it based on the following
personal criteria:

1) It's working on a production site, so I have a fair amount of confidence it works.

2) It lives in its own package, so it's ability to do harm to anything else is low.

3) I'd like to deprecate requiredif in favor of this because it has a much simpler
syntax and is much ore powerful, and I fear that if a lot of people start using
requiredif after the 1.1 release, it will be hard to get them off the habbit.
How do you deprecate a validation function, anyway?

4) I've had a number of direct requests for the code, which tends to make me
believe it is of general utility.

James