Issue number:  13748
when Issue.addComment() is called, email will get sent as part of the
business object instead of as part of the action. of course we will want
to go through the entire app and move the email sending from the actions
into the business logic, but for now, this is the only requirement where
we need this functionality.