Localize the charset for all error handlers.
Also change to always set the charset if it is available, since IMHO it is better to send a charset then to assume that the browser defaults to iso-latin-1 (which is false for MSIE, for example).

This clearly needs to be done for Japanese servers.  It should be a pretty minor change for en, fr, and sp servers.  Other Locals where probably just as broken before.

Fix for bug #20361.
Reported By: Gary gbraven@yahoo.com