for issue 19170: * MITList.java: getPotentialAssignees() : if includeCommitters flag is true, add ISSUE_ENTER permission to permission list used to retrieve users. * ScarabModule.java: getUsers(): pass includeCommitters flag to MitList.getPotential Assignees() * IssueSearch.java: getUserAttributes(): pass false as activeOnly argument so that the search uses inactive user attributes. * GlobalMacros.vm: #userList(); UserList.vm: if there are no user attributes to assign users to, still show screen, since committed by users are eligible to be searched on. * reports/SelectUsers.vm: passing includeCommitters flag to #userList macro as true
for issue 19241:
* MitList.java: getCommonRModuleUserAttributes():  new logic for saving issue
list attributes for cross issuetype searches.  first check saved RMUAs and see
if there are any common ones for these issue types. if so, go with those. if
not, delete the RMUAs and check for common attributes in the saved RMUAs for the
first issue type in the list, and check the default RMUAs for the first issue
type in the list. If you still do not have at least 3, go through common
attributes until you come up with 3. Save the resulting attributes as RMUAs.
* RModuleUserAttribute.java: delete(): make sure to specify list id in criteria
when deleting a RMUA.
Issue number: 19241, 19170
Reviewed by:   19241: venkat; 19170: kmaples