This is quite a massive commit. Here is a short summary:
- The o.a.l.helpers.PatternParser class has been moved to
 o.a.l.pattern package. 

Many of the small PatternConverter classes contained in 
PatternParser.java  file have been taken out  and placed in 
o.a.l.pattern package as  independent class files.

The Layout class has changed. A new method 

  void format(Writer output, LoggingEvent event)

has been added. Instead of returning a String, Layouts now write 
directly to the output stream. This significantly improves performance, 
in particular by avoiding the  creation of many temporary strings.

The  method 

  String  format(LoggingEvent event) 

in Layout is kept around for compatibility. This method is now 
implemented in terms of the format(Writer, LoggingEvent) variant 
where Writer is a CharArrayWriter.

The changes pass the existing unit tests.