The max length javascript validation considers whitespace as a character.  The server side max length check does *not* consider whitespace when comparing the number of characters to the maximum allowed.  The javascript is correct so revert validateMaxLength() logic to how it was in revision 1.6.
Whitespace *should* count against the total number of characters in a string.
An example is a database field defined as CHAR(2).  The maxlength validation should
limit the input to 2 characters regardless of their type.  If it didn't, and I input 3 space
characters the validation would pass and the database would throw an error.