* src/java/org/tigris/scarab/actions/ModifyIssue.java
doMove: if condition is added in this method. When the project owner
permission is revoked for the logged in user, the size of the list that
contains permitted modules, will become zero. Based on this condition, we
decide whether to take the user to the next page or retain him in the current
page along with alert message.

Issue: PCN19606
Submitted by: irk@collab.net
Reviewed by: Elicia