This fixes issue SCB984, but it also fixes a wider problem. Resource bundles come in the form:
  ResourceBundle.properties
  ResourceBundle_<language code>.properties
  ResourceBundle_<language code>_<country code>.properties

Each bundle has a higher priority than the previous one. So,
if a bundle can be found that matches both the language *and*
country settings (the third option), it will be used.
Otherwise, a resource bundle that explicitly matches just
the language setting (the second option) will be used. Finally,
if all else fails, the default bundle is used (the first
option).

OK, now that I've summarised how resource bundles work, on
to the issue. The ScarabBundle had all three of the above
options available. However, the InitDbScarabBundle only had
instances of the third option. This meant that Scarab would
always fail on a locale not explicitly supported.

So, since it seems to be the job of the build file to create
the first and second option resource bundles, I have modified
it to generate the required files for the InitDbScarabBundle.
I have also added the chinese resource bundles to the build.

Files modified:

  build/build.xml