Fix for issue SCB1164:
  http://scarab.collab.net/scarab/issues/id/SCB1164

The method IssueType.getAllRIssueTypeOptions() was returning null
if an option list attribute had no options. Unfortunately, some
code further down the line was assuming (reasonably in my view)
that an empty list would be returned in such a case.

So, the fix just involves the method returning an empty ArrayList
when no options are found.

Files modified:

  src/java/org/tigris/scarab/om/IssueType.java