Fixes for SCB544 and SCB670:
 http://scarab.collab.net/scarab/issues/id/SCB544
 http://scarab.collab.net/scarab/issues/id/SCB670

Dates can now be entered in the user's own locale-specific format,
while the example date pattern displayed on the advanced query
screen is now localised as well. Since the pattern for display can
be different to the pattern used to parse dates, there are two new
entries per date type in the localisation resource bundle, i.e. two
for a date and two for a date/time.

Moving on to the second issue, although the results per page
attribute was being saved with queries, the issue list page was
not picking up that saved value. Running a saved query now adds
the required parameters to the RunData so that the issue list
page can find the results per page, sort column and sort polarity.


Files modified:

 src/conf/classes/ScarabBundle_en.properties
 src/conf/classes/ScarabBundle_de.properties
 src/conf/classes/ScarabBundle_fr.properties
 src/java/org/tigris/scarab/actions/Search.java
 src/java/org/tigris/scarab/actions/base/RequireLoginFirstAction.java
 src/java/org/tigris/scarab/actions/base/ScarabTemplateAction.java
 src/java/org/tigris/scarab/tools/ScarabLocalizationTool.java
 src/java/org/tigris/scarab/tools/ScarabRequestTool.java
 src/java/org/tigris/scarab/util/word/IssueSearch.java
 src/webapp/WEB-INF/templates/macros/AdvancedQueryMacro.vm
 src/webapp/WEB-INF/templates/macros/GlobalMacros.vm
 src/webapp/WEB-INF/templates/screens/IssueList.vm
 src/webapp/WEB-INF/templates/screens/reports/ConfineDataset.vm
 xdocs/changes.xml