Fixes issue SCB1131:
 http://scarab.collab.net/scarab/issues/id/SCB1131

The "possible duplicates found" screen (Wizard2.vm) was including
the date that issues were closed. Unfortuantely this relied on the
status attribute existing in the database, with well-defined options
on it. In particular, the code required the status attribute to have
a "closed" value.

I have removed the method "getClosedDate" from Issue, and the screen
that was using the feature has had all its references removed.
 

Files modified:

 src/java/org/tigris/scarab/om/Issue.java
 src/webapp/WEB-INF/templates/screens/entry/Wizard2.vm
 xdocs/changes.xml