This change adds a Web start specific build feature.  If you don't use the webstart target, this build file works as expected.
The Web start build excludes the JMSReceiver + DBReceiver from the core jar, and places it in a separate
one.   IT also prompts you for the certificate to sign with, which I use and specified in my local build.properties file,
but then this could be used if anyone else in the team needed to sign the jar with their own.