Add system integration tests to validate the values returned by property getters on FacesContext, ExternalContext, and HttpServletRequest.
The current test is based on the use of a wrapper around the servlet request,
that was introduced because of bug report #29809.  It is not yet clear what
the final resolution of this issue is, but it may involve further refinements
in the functionality of the Struts-Faces integration library before a 1.0
release.