New download page, hooked up to the mirrors, and loaded up for the Struts 1.2.2 release. This is based on the HTTPD download page, and uses the more elegant of the two mirror page options.
This XML file generates an HTML document that the <xmlvalidate> task does
not like. Some of the problems could probably be fixed, but in XHTML 1.0
(which is what we are validating against), the <img> element cannot have
an 'align' attribute (which is what keeps the mirror's logo over on the
right). There's probably an alternative way of making this happen, but in
the interests of getting this into CVS, I've simply excluded this file
from the XHTML validation check.