Failed form validations were resulting in ActionErrors objects which were never being saved anywhere!
I solved this with the possibly contentious solution of saving them directly
in the validate() method.  I did this mostly to preserve the separation of
servlet API specifics (the errors are saved in the HttpServletRequest scope).

The only other alternative would be to add another abstract method to
AbstractValidateActionForm, such as "saveErrors" but that seems kind of
pedantic.

Would it clear things up more to have validate return boolean and clearly give
it the responsibility for "saving" any errors it finds?