This is a merge of change "2005-01-13 15:00" from the carol_1_8_5_br branch back onto the trunk.  The merge includes the following changes from the branch:
 build.xml (1.27.2.3)
 test/build.xml (1.11.2.1)
 test/conform/carol_all_tests.xml (1.12.2.2)
 test/conform/generic/carol_double.xml (1.11.2.2)
 test/ext/java.policy (1.3.2.1)
 test/src/org/objectweb/carol/jtests/conform/basic/clients/MultiProtocolTests.java (1.5.2.1)
 test/src/org/objectweb/carol/jtests/conform/basic/server/BasicServer.java (1.6.2.1)
 test/src/org/objectweb/carol/jtests/conform/util/ProcessRunner.java (1.1.2.1)
 test/src/org/objectweb/carol/jtests/conform/util/ProcessStopper.java (1.1.2.1)

This change is described in
  http://mail-archive.objectweb.org/carol/2005-01/msg00000.html

The original log message can be viewed here:
http://cvs.forge.objectweb.org/cgi-bin/viewcvs.cgi/carol/carol/test/conform/generic/carol_double.xml?rev=1.11.2.2&only_with_tag=carol_1_8_5_br&content-type=text/vnd.viewcvs-markup

The original log message is also reproduced below for convenience

..............................................................................

The purpose of this commit is to change the way we set up and tear
down multiple concurrent processes when running unit tests.

For background information, see these threads:

 - http://mail-archive.objectweb.org/carol/2004-12/msg00005.html
 - http://mail-archive.objectweb.org/carol/2004-12/msg00022.html

The unit tests verify correctness of various RMI implementations
provided by CAROL.  To test remote method calls, we normally need
three different processes running concurrently:

  (a) registry;
  (b) remote server;
  (c) remote client.

(In actuality, this ranges from 2 to 5 concurrent processes under
different scenarios that CAROL currently tests for.)

These processes should be started in the above order and shut down in
the reverse order.  Prior to this commit, CAROL had no _precise_ way
of starting these processes in the right order and tearing them down
correctly.  The desired effect was achieved by using Ant's <parallel>
task in concert with <sleep> to start the processes in the right
order.  For teardown, we were relying on the 'timeout' attribute of
the <java> task.  One of the side effects of this approach was that
the tests took 9 minutes to run regardless of how fast your computer
was.

This commit eliminates the use of sleeps and timeouts.  Instead, we
now rely on the Ant's <waitfor> task to ensure that the required
processes are started in the right order.  Two new helper classes are
introduced to allow us to programmatically shut down the processes by
sending a command over TCP.  These two new classes are
  org.objectweb.carol.jtests.conform.util.ProcessRunner
  org.objectweb.carol.jtests.conform.util.ProcessStopper

The above accounts for the bulk of the changes committed in this
checkin.

Another small change was to _not_ generate javadocs as part of running
tests.

The end result is that tests now take anywhere from 1 to 3 minutes to
finish depending on how fast your machine is.  This is down from 9
minutes.

This commit introduces a seeming regression.  Tests involving either
jrmp1.1 or jrmp1.2 now fail intermittently.  For tests that do not
involve either of these two protocols, there are no regressions.  I
will investigate intermittent failures and commit a fix as soon as I
have it.

The following files are part of this commit:

  build.xml
  test/build.xml
  test/conform/carol_all_tests.xml
  test/conform/generic/carol_double.xml
  test/ext/java.policy
  test/src/org/objectweb/carol/jtests/conform/basic/clients/MultiProtocolTests.java
  test/src/org/objectweb/carol/jtests/conform/basic/server/BasicServer.java
  test/src/org/objectweb/carol/jtests/conform/util/ProcessRunner.java
  test/src/org/objectweb/carol/jtests/conform/util/ProcessStopper.java