This is a merge of change "2005-01-28 16:08" from the carol_1_8_5_br branch back onto the trunk.  It copies the following file:
  src/org/objectweb/carol/jndi/ns/IIOPCosNaming.java (1.6.2.1)

This change is described in
  http://mail-archive.objectweb.org/carol/2005-01/msg00011.html

The original log message can be viewed here:
http://cvs.forge.objectweb.org/cgi-bin/viewcvs.cgi/carol/carol/src/org/objectweb/carol/jndi/ns/IIOPCosNaming.java?rev=1.6.2.1&content-type=text/vnd.viewcvs-markup

For convenience, the original log message is reproduced below:

..............................................................................

The method isStarted() checks whether a name server is already
running.  The check is performed by attempting to instantiate a new
InitialContext.  This instantiation attempt will raise an exception,
if the name server is not running.  The previous implementation of
this idea had the following flaw.

Under Sun's JVM, attempting to instantiate a new InitialContext in the
absence of a running name server results in a
javax.naming.CommunicationException.  The previous implementation of
isStarted() would catch this exception and correctly return "false".

This doesn't work under IBM's JDK.  When no name server is running,
instantiating a new InitialContext results in a
javax.naming.ConfigurationException under IBM's JDK.  Under the
previous implementation of isStarted(), this case would follow a
different execution path.  The ConfigurationException would be caught
under the catch-all clause "catch (Exception e)" and the method would
return "true", _erroneously_ indicating that the name server *is*
already running.

This patch introduces two changes:

 (a) it removes the highly questionable catch-all clause
     "catch (Exception e)"

 (b) instead of catching javax.naming.CommunicationException, we now
     catch javax.naming.NamingException.

     This is correct because both CommunicationException thrown by Sun
     and ConfigurationException thrown by IBM are subclasses of
     NamingException.  As a matter of fact, the constructor
     InitialContext(Properties) is documented to throw
     NamingException.  Catching a specific subclass of NamingException
     is highly non-portable.