This is a merge of change "2005-01-31 17:02" from the carol_1_8_5_br branch back onto the trunk.  The following changes are applied:
  test/conform/generic/carol_double.xml (1.11.2.3),
  test/src/org/objectweb/carol/jtests/conform/util/ProcessRunner.java (1.1.2.2)

The original log message is reproduced here for convenience:

..............................................................................
This change affects the following two files:

  test/conform/generic/carol_double.xml
  test/src/org/objectweb/carol/jtests/conform/util/ProcessRunner.java

It fixes intermittent test failures described in this commit message:

http://cvs.forge.objectweb.org/cgi-bin/viewcvs.cgi/carol/carol/test/conform/generic/carol_double.xml?rev=1.11.2.2&content-type=text/vnd.viewcvs-markup

For more background info, see also
http://mail-archive.objectweb.org/carol/2005-01/msg00000.html

These intermittent failures were occurring due to the following
reasons.

There are currently two kinds of tests in CAROL: "single" and
"multi".  In "single" tests, three procecess are started: a
registry, the server, and the client.  In "multi" tests, five
processes are started concurrently: two registries, one remote server,
and two remote clients.

Only "multi" tests were experiencing intermittent failures.  These
failures were occurring due to the following asymmetry between the two
clients.

In order to successfully execute "multi" tests, the aforementioned
five processes have to be started and stopped in the correct order.
The startup order is as follows:

 (1) the two registries must be started;

 (2) once the registies are up and running, the remote server must be
     started;

 (3) once the server is running, the two clients must be started.

In our case, the clients are JUnit processes that execute and test
remote method calls.  The clients terminate on their own without
requiring any special external intervention.  In contrast, the server
and the registry must be explicitly shut down by the test harness.

The shutdown must occur in the reverse order.

 (1) once _both_ clients are finished running, the server must be
     shutdown.

 (2) once the server is down, the registries must be shut down.

The bug causing intermittent failures was present in carol_double.xml.
The problem was that we were shutting down the server as soon as the
first client was finished.  Depending on whether or not the second
concurrently executing client was still running at the time of remote
server shutdown, intermittent failures might or might not occur.

The fix is to make sure that the remote server is only shut down after
_both_ clients have exited.