Removed the jonathan.jvmarg property.  It appears to have been copied and pasted from test/genstub/rmic_jonathan.xml into build.xml.  In order to get expanded correctly, this property depended on the following values having been set:
    test.externals
    test.build
    test.ext

These are not set in build.xml.  So, the property ended up being to
set to a garbage string with unexpanded $ signs in it:

  -Xbootclasspath/p:${test.externals}/jonathan.jar:...

Removing the jonathan.jvmarg altogether did not cause any problems in
build.xml.  Encouraged by this, I tried removing it from
test/genstub/rmic_jonathan.xml as well.  No problems there either.