Cleaned out the test/externals/ directory as discussed in http://mail-archive.objectweb.org/carol/2005-02/msg00018.html
The following .jar files are removed because their exact duplicates
exist in the externals/ directory:

  test/externals/jonathan-core.jar
  test/externals/kilim.jar
  test/externals/commons-cli.jar
  test/externals/jeremie.jar

The following .jar files are removed because of their obsolescence:
  test/externals/avalon-framework.jar
  test/externals/jacorb.jar
  test/externals/junit.jar
  test/externals/logkit.jar

Specifically, test/externals/jacorb.jar is a really old version of
jacorb.jar (the same as externals/jacorb.jar revision 1.1).
avalon-framework.jar and logkit.jar were used by the old version of
jacorb.jar.  Similarly, junit.jar is ancient and not currently used.

The following .jar files are moved from test/externals/ to externals/

  kilim-tools.jar
  log4j.jar
  nanoxml-lite-2.2.1.jar
  ow_util_log_file.jar
  ow_util_log_wrp_log4j.jar

To keep build files working, all references to the "test.externals"
property are replaced with references to the "externals" property.
The following files are affected:

  build.xml
  test/build.xml
  test/conform/generic/carol_double.xml
  test/genstub/rmic_jonathan.xml
  test/genstub/rmic_sun.xml