Minor cleanup.
Prior to these change, build.xml used two different properties for the
same purpose: "version" and "carol.version".  The former was defined
in config/configure.properties and hasn't been updated in ages.  (It
was still set to 1.5.1.)  The latter is defined in build.xml.

I removed "version" and updated affected targets to use
"carol.version" instead.