Removed the target "build.test" from both build.xml and test/build.xml.
This target didn't really belong in build.xml.  It should be
sufficient for build.xml to call the target "run.tests" in
test/build.xml without also having to explicitly call "build.test".
It is then up to test/build.xml to decide whether or not to build the
tests.

In test/build.xml, the target "build.test" had a trivial
implementation.  By inlining it, we reduce the visual structural
complexity of the build process.

Compare
http://cvs.forge.objectweb.org/cgi-bin/viewcvs.cgi/carol/carol/doc/usermanual/img/test-build.png?rev=1.2&content-type=text/vnd.viewcvs-markup
and
http://cvs.forge.objectweb.org/cgi-bin/viewcvs.cgi/carol/carol/doc/usermanual/img/test-build.png?rev=1.3&content-type=text/vnd.viewcvs-markup


Also compare

http://cvs.forge.objectweb.org/cgi-bin/viewcvs.cgi/carol/carol/doc/usermanual/img/build.png?rev=1.1&content-type=text/vnd.viewcvs-markup
and
http://cvs.forge.objectweb.org/cgi-bin/viewcvs.cgi/carol/carol/doc/usermanual/img/build.png?rev=1.2&content-type=text/vnd.viewcvs-markup