This commit mostly accomplishes the goal set forth in the log message on commit "2005-02-08 13:13", viewable here: http://cvs.forge.objectweb.org/cgi-bin/viewcvs.cgi/carol/carol/test/build.xml?rev=1.17&content-type=text/vnd.viewcvs-markup As a result of this commit, we now have a single config file per each protocol.
    test/ext/iiop.properties
    test/ext/jeremie.properties
    test/ext/jrmp.properties

These files are used to dynamically construct 28 different
client-server configurations that we currently test.

This number breaks down as follows.

We have 8 single-protocol tests: each of the four supported protocols
is tested in two configurations -- with and without an externally
started name server.

We also test 10 different two-protocol configurations.  Each such
configuration is tested twice: with and without an externally started
name server.  That makes for 20 tests.

In total, we test 8 single-protocol configurations and 20 two-protocol
configurations.

We used to have numerous config files for these test cases.  These
files are removed in this commit.

The next step is to eliminate a hard-coded dependency on
com.sun.corba.se.internal.CosNaming.TransientNameServer that still
exists in test/ext/iiop.properties.  This is now easily achievable,
because this dependency only exists in one file now.  Prior to these
commit, it existed in multiple files.