Add two new methods:    public static void pushSubjectContext(Subject subject,       Principal principal, Object credential)    public static SubjectContext popSubjectContext()
These effectively replace the use of setPrincipal, setCredential, setSubject when running in the server. The setPrincipal, setCredential still have some use in client vms that are not running in multi-threaded mode(server=false). The setSubject has no use as it is not used on the client to propagate security context.