Removed the useless private static field "init".
Since the "init" field is private, we know it wasn't used anywhere
outside the JInterceptorStore class.  Inside the class, it was
referenced three times.  The following pared down skeletal version of
the class shows all of the three occurrences:

 | public class JInterceptorStore {
 |
 |     private static boolean init = false;  // 1st mention
 |
 |     static {
 |         if (!init) {                      // 2nd mention
 |             try {
 |                 // do stuff that may throw an exception
 |                 init = true;              // 3rd mention
 |             } catch (Exception e) {
 |                 TraceCarol.error("JrmpPRODelegate(), No interceptors found", e);
 |             }
 |         }
 |
 |     }
 |
 |     // init is never referenced again below
 | }

The "if (!init)" test was obviously redundant, since we know that
"init" was false at that point, therefore "!init" was true, and the
condition's body was _always_ executed.

Resetting the "init" to true inside the try-catch statement was
redundant, because the value of "init" was never referenced after
that.

Because of all that, the field was safe to remove.