Made the getJRMPInitializers() method thread-safe.
The method wasn't thread-safe for the following reasons.

It lazily instantiates the static "initializers" array. The pared down
outline of this method looks like so:

    public static String[] getJRMPInitializers() {
        // if exists, return it
        if (initializers != null) {
            return initializers;
        }
        // else, compute it.
        // ...
        initializers = new String[sz];
        for (int i = 0; i < sz; i++) {
            initializers[i] = (String) ins.get(i);
        }
        return initializers;
    }

Suppose Thread A comes along and finds the "initializers" field to be
null.  So, Thread A executes

       initializers = new String[sz];

and begins populating the newly allocated array.

Suppose further that while Thread A is still populating the
"initializers" array, Thread B comes along.  It finds "initializers"
to be non-null.  So, in Thread B, this method returns a partially
initialized value of this field.  Not good.